requirements.txt
python-dotenv
pandas
numpy
pymongo
pymongo[srv]
certifi
scikit-learn
dill
pyaml
# -e .